Move things from the main menu to their respective submenus
